service: move redis event streaming to its own type
